1.20.1 forge

+++ src/main/java/net/acetheeldritchking/cataclysm_spellbooks/spells/ender/VoidRuneSpell.java
@@ -99,7 +99,7 @@ public class VoidRuneSpell extends AbstractSpell {
                 double targetY = targetEntity.getY();
                 double targetZ = targetEntity.getZ();
 
-                BlockPos blockPos = new BlockPos(targetX, targetY, targetZ);
+                BlockPos blockPos = new BlockPos((int) targetX, (int) targetY, (int) targetZ);
 
                 for (int i = 0; i < getDuration(spellLevel, entity); i++)
                 {
